base_requirements.txt
ansible>=1.2
Flask-Script
nose
nose-testconfig
requests>=1.2
URLObject
